refactor: extract the missing media pipeline seam
Mirror the missingModelPipeline seam so the media pipeline is reachable
outside ComfyApp. Pure move: no behavior change.
